feat: refactoring - further refact and rewrite. Focus on Server-side
