+++ src/client/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>CASPARCG CLIPTOOL</title>
+  </head>
+  <body style="margin:0;padding:0">
+    <div id="root">
+    </div>
+    <script src="index.js"></script>
+  </body>
+</html>
