+++ src/model/ccgDatasctructure.ts
