+++ src/model/reducers/appNavReducer.ts
