+++ src/server/global.ts
@@ -1,10 +1,9 @@
-declare module NodeJS {
+export declare module NodeJS {
     interface Global {
         mediaFolders: string
         dataFolders: string
         templateFolders: string
         serverVersion: string
-        graphQlServer: any
         oscServer: any
     }
 }
